GETP-216 rename: common 컴포넌트의 util 클래스를 util 컴포넌트로 분리
