+++ src/main/java/es/princip/getp/application/auth/service/VerificationService.java
@@ -2,11 +2,14 @@ package es.princip.getp.application.auth.service;
 
 import es.princip.getp.application.auth.exception.IncorrectVerificationCodeException;
 import es.princip.getp.application.auth.exception.NotFoundVerificationException;
-import es.princip.getp.common.util.RandomUtil;
 import es.princip.getp.domain.common.model.Email;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+
+import java.security.NoSuchAlgorithmException;
+import java.security.SecureRandom;
+import java.util.Random;
 
 @Service
 @Transactional(readOnly = true)
@@ -33,7 +36,7 @@ public class VerificationService {
     @Transactional
     public void sendVerificationCode(Email email) {
         verificationRepository.deleteById(email.getValue());
-        final String verificationCode = RandomUtil.generateRandomCode(VERIFICATION_CODE_LENGTH);
+        final String verificationCode = generateRandomCode(VERIFICATION_CODE_LENGTH);
         verificationSender.send(email, verificationCode);
         final EmailVerification emailVerification = new EmailVerification(email.getValue(), verificationCode, EXPIRATION_TIME);
         verificationRepository.save(emailVerification);
@@ -47,5 +50,18 @@ public class VerificationService {
             throw new IncorrectVerificationCodeException();
         }
         verificationRepository.delete(verification);
+    }
+
+    private static String generateRandomCode(int length) {
+        try {
+            final Random random = SecureRandom.getInstanceStrong();
+            final StringBuilder builder = new StringBuilder();
+            for (int i = 0; i < length; i++) {
+                builder.append(random.nextInt(10));
+            }
+            return builder.toString();
+        } catch (NoSuchAlgorithmException exception) {
+            throw new RuntimeException();
+        }
     }
 }
