--- src/main/java/es/princip/getp/common/util/StringUtil.java
@@ -1,15 +1,0 @@
-package es.princip.getp.common.util;
-
-public class StringUtil {
-
-    public static String camelToSnake(String camelCase) {
-        if (camelCase == null || camelCase.isEmpty()) {
-            return camelCase;
-        }
-
-        String regex = "([a-z])([A-Z]+)";
-        String replacement = "$1_$2";
-
-        return camelCase.replaceAll(regex, replacement);
-    }
-}
